+++ tests/unit/test__helpers.py
@@ -24,11 +24,6 @@ try:
 except ImportError:  # pragma: NO COVER
     bigquery_storage = None
 
-try:
-    import pyarrow
-except ImportError:  # pragma: NO COVER
-    pyarrow = None
-
 
 @unittest.skipIf(bigquery_storage is None, "Requires `google-cloud-bigquery-storage`")
 class TestBQStorageVersions(unittest.TestCase):
@@ -98,64 +93,6 @@ class TestBQStorageVersions(unittest.TestCase):
         versions = self._object_under_test()
         with mock.patch("google.cloud.bigquery_storage.__version__", new="2.5.0"):
             assert not versions.is_read_session_optional
-
-
-@unittest.skipIf(pyarrow is None, "Requires `pyarrow`")
-class TestPyarrowVersions(unittest.TestCase):
-    def tearDown(self):
-        from google.cloud.bigquery import _helpers
-
-        # Reset any cached versions since it may not match reality.
-        _helpers.PYARROW_VERSIONS._installed_version = None
-
-    def _object_under_test(self):
-        from google.cloud.bigquery import _helpers
-
-        return _helpers.PyarrowVersions()
-
-    def _call_try_import(self, **kwargs):
-        from google.cloud.bigquery import _helpers
-
-        _helpers.PYARROW_VERSIONS._installed_version = None
-        return _helpers.PYARROW_VERSIONS.try_import(**kwargs)
-
-    def test_try_import_raises_no_error_w_recent_pyarrow(self):
-        from google.cloud.bigquery.exceptions import LegacyPyarrowError
-
-        with mock.patch("pyarrow.__version__", new="5.0.0"):
-            try:
-                pyarrow = self._call_try_import(raise_if_error=True)
-                self.assertIsNotNone(pyarrow)
-            except LegacyPyarrowError:  # pragma: NO COVER
-                self.fail("Legacy error raised with a non-legacy dependency version.")
-
-    def test_try_import_returns_none_w_legacy_pyarrow(self):
-        with mock.patch("pyarrow.__version__", new="2.0.0"):
-            pyarrow = self._call_try_import()
-            self.assertIsNone(pyarrow)
-
-    def test_try_import_raises_error_w_legacy_pyarrow(self):
-        from google.cloud.bigquery.exceptions import LegacyPyarrowError
-
-        with mock.patch("pyarrow.__version__", new="2.0.0"):
-            with self.assertRaises(LegacyPyarrowError):
-                self._call_try_import(raise_if_error=True)
-
-    def test_installed_version_returns_cached(self):
-        versions = self._object_under_test()
-        versions._installed_version = object()
-        assert versions.installed_version is versions._installed_version
-
-    def test_installed_version_returns_parsed_version(self):
-        versions = self._object_under_test()
-
-        with mock.patch("pyarrow.__version__", new="1.2.3"):
-            version = versions.installed_version
-
-        assert version.major == 1
-        assert version.minor == 2
-        assert version.micro == 3
-
 
 class Test_not_null(unittest.TestCase):
     def _call_fut(self, value, field):
