+++ tests/unit/test__pandas_helpers.py
@@ -41,10 +41,11 @@ from google import api_core
 
 from google.cloud.bigquery import exceptions
 from google.cloud.bigquery import _helpers
+from google.cloud.bigquery import _pyarrow_helpers
 from google.cloud.bigquery import schema
 from google.cloud.bigquery._pandas_helpers import _BIGNUMERIC_SUPPORT
 
-pyarrow = _helpers.PYARROW_VERSIONS.try_import()
+pyarrow = _pyarrow_helpers.PYARROW_VERSIONS.try_import()
 
 if pyarrow:
     import pyarrow.parquet
@@ -346,14 +347,14 @@ def test_bq_to_arrow_data_type_w_struct(module_under_test, bq_type):
         pyarrow.field("field04", pyarrow.int64()),
         pyarrow.field("field05", pyarrow.float64()),
         pyarrow.field("field06", pyarrow.float64()),
-        pyarrow.field("field07", _helpers.pyarrow_numeric()),
-        pyarrow.field("field08", _helpers.pyarrow_bignumeric()),
+        pyarrow.field("field07", _pyarrow_helpers.pyarrow_numeric()),
+        pyarrow.field("field08", _pyarrow_helpers.pyarrow_bignumeric()),
         pyarrow.field("field09", pyarrow.bool_()),
         pyarrow.field("field10", pyarrow.bool_()),
-        pyarrow.field("field11", _helpers.pyarrow_timestamp()),
+        pyarrow.field("field11", _pyarrow_helpers.pyarrow_timestamp()),
         pyarrow.field("field12", pyarrow.date32()),
-        pyarrow.field("field13", _helpers.pyarrow_time()),
-        pyarrow.field("field14", _helpers.pyarrow_datetime()),
+        pyarrow.field("field13", _pyarrow_helpers.pyarrow_time()),
+        pyarrow.field("field14", _pyarrow_helpers.pyarrow_datetime()),
         pyarrow.field("field15", pyarrow.string()),
     )
     expected = pyarrow.struct(expected)
@@ -394,14 +395,14 @@ def test_bq_to_arrow_data_type_w_array_struct(module_under_test, bq_type):
         pyarrow.field("field04", pyarrow.int64()),
         pyarrow.field("field05", pyarrow.float64()),
         pyarrow.field("field06", pyarrow.float64()),
-        pyarrow.field("field07", _helpers.pyarrow_numeric()),
-        pyarrow.field("field08", _helpers.pyarrow_bignumeric()),
+        pyarrow.field("field07", _pyarrow_helpers.pyarrow_numeric()),
+        pyarrow.field("field08", _pyarrow_helpers.pyarrow_bignumeric()),
         pyarrow.field("field09", pyarrow.bool_()),
         pyarrow.field("field10", pyarrow.bool_()),
-        pyarrow.field("field11", _helpers.pyarrow_timestamp()),
+        pyarrow.field("field11", _pyarrow_helpers.pyarrow_timestamp()),
         pyarrow.field("field12", pyarrow.date32()),
-        pyarrow.field("field13", _helpers.pyarrow_time()),
-        pyarrow.field("field14", _helpers.pyarrow_datetime()),
+        pyarrow.field("field13", _pyarrow_helpers.pyarrow_time()),
+        pyarrow.field("field14", _pyarrow_helpers.pyarrow_datetime()),
         pyarrow.field("field15", pyarrow.string()),
     )
     expected_value_type = pyarrow.struct(expected)
@@ -1117,7 +1118,7 @@ def test_dataframe_to_parquet_without_pyarrow(module_under_test, monkeypatch):
     mock_pyarrow_import.side_effect = exceptions.LegacyPyarrowError(
         "pyarrow not installed"
     )
-    monkeypatch.setattr(_helpers.PYARROW_VERSIONS, "try_import", mock_pyarrow_import)
+    monkeypatch.setattr(_pyarrow_helpers.PYARROW_VERSIONS, "try_import", mock_pyarrow_import)
 
     with pytest.raises(exceptions.LegacyPyarrowError):
         module_under_test.dataframe_to_parquet(pandas.DataFrame(), (), None)
